Relocate PKE secret code
